Langgraph/OCR_AI_Agent/pyproject.toml
[project]
name = "ocr-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-azure-ai>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.2",
    "langchain-groq>=1.0.0",
    "langchain-openai>=1.0.1",
    "langgraph>=1.0.2",
    "numpy>=2.3.4",
    "opencv-python>=4.11.0.86",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
]
